Decorators for image and attachment
