--- app/models/spina/attachment.rb
@@ -1,39 +1,0 @@
-module Spina
-  class Attachment < ApplicationRecord
-    has_one_attached :file
-
-    has_one :page_part, as: :page_partable
-    has_many :structure_parts, as: :structure_partable
-    has_and_belongs_to_many :attachment_collections, join_table: 'spina_attachment_collections_attachments'
-
-    attr_accessor :_destroy
-
-    scope :sorted, -> { order('file ASC') }
-
-    def name
-      file.filename.to_s
-    end
-
-    def content
-      file if file.attached?
-    end
-
-    alias_method :old_update, :update
-    def update(attributes)
-      if attributes["_destroy"] == "1" && attributes["file"].blank?
-        self.page_part.destroy
-      else
-        old_update(attributes)
-      end
-    end
-
-    def convert_to_json!
-      attachment = Spina::Parts::Attachment.new
-      attachment.attachment_id = id
-      attachment.signed_blob_id = file&.blob&.signed_id
-      attachment.filename = name
-      attachment
-    end
-
-  end
-end
