+++ lib/spina/upgrade/decorators/image_decorator.rb
@@ -1,0 +1,15 @@
+module Spina
+  module Upgrade
+    module ImageDecorator
+
+      def convert_to_json!
+        image = Spina::Parts::Image.new
+        image.image_id = id
+        image.signed_blob_id = file&.blob&.signed_id
+        image.filename = name
+        image
+      end
+      
+    end
+  end
+end
